Refactor validation middleware
